+++ packages/zowe-explorer-api/src/index.ts
@@ -9,16 +9,21 @@
  *
  */
 
+export * from "./globals/Gui";
+export * from "./logger/IZoweLogger";
 export * from "./profiles/UserSettings";
 export * from "./profiles/ProfilesCache";
 export * from "./profiles/ZoweExplorerApi";
 export * from "./profiles/ZoweExplorerZosmfApi";
-export * from "./logger/IZoweLogger";
+export * from "./security/KeytarApi";
+export * from "./security/KeytarCredentialManager";
+export * from "./security/KeytarApi";
+export * from "./security/KeytarCredentialManager";
 export * from "./tree/ZoweExplorerTreeApi";
 export * from "./tree/ZoweTreeNode";
 export * from "./tree/IZoweTree";
 export * from "./tree/IZoweTreeNode";
+export * from "./utils";
 export * from "./vscode/ZoweVsCodeExtension";
-export * from "./security/KeytarApi";
-export * from "./security/KeytarCredentialManager";
-export * from "./globals/Gui";
+export * from "./vscode/ui";
+export { imperative } from "@zowe/cli";